requirements.txt
absl-py==2.1.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.24
dnspython==2.7.0
etils==1.9.4
eventlet==0.38.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
fonttools==4.54.1
fsspec==2024.9.0
glfw==2.7.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
ikpy==3.4.2
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
labmaze==1.0.6
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mjcf_urdf_simple_converter==0.4
mpmath==1.3.0
mujoco==3.2.3
netifaces==0.10.6
numpy==2.1.1
numpy-stl==3.1.2
opencv-python==4.10.0.84
packaging==24.1
pillow==11.0.0
protobuf==5.28.3
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
python-utils==3.9.0
requests==2.32.3
rich==13.9.3
scipy==1.14.1
simple-websocket==1.1.0
six==1.16.0
sympy==1.13.3
tk==0.1.0
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.20.2
